Add exposure time as parameter and fix is_bigendian
The exposure time of the PCO camera is now not hard coded anymore
and can be set via ros parameter for example with an yaml file.

The node caused rqt to crash due to a wrong setting of
is_bigendian. This problem was fixed by setting the is_bigendian
always to 0.